TASK-4: Modified some minor things, refactored code, created directories
