+++ TASK-4/Interfaces/IUserInfo.cs
@@ -1,4 +1,4 @@
-namespace Persons;
+namespace TASK_4.Interfaces;
 
 public interface IUserInfo
 {
