+++ TASK-4/Persons/Employee.cs
@@ -1,4 +1,6 @@
-namespace Persons;
+using TASK_4.Interfaces;
+
+namespace TASK_4.Persons;
 
 public class Employee : User, IUserInfo
 {
@@ -15,31 +17,37 @@ public class Employee : User, IUserInfo
         get { return _position; }
         set { _position = value; }
     }
+    
     public string PositionDescription
     {
         get { return _positionDescription; }
         set { _positionDescription = value; }
     }
+    
     public double Salary
     {
         get { return _salary; }
         set { _salary = value; }
     }
+    
     public string CompanyName
     {
         get { return _companyName; }
         set { _companyName = value; }
     }
+    
     public string CompanyCountry
     {
         get { return _companyCountry; }
         set { _companyCountry = value; }
     }
+    
     public string CompanyCity
     {
         get { return _companyCity; }
         set { _companyCity = value; }
     }
+    
     public string CompanyAddress
     {
         get { return _companyAddress; }
