+++ TASK-4/Program.cs
@@ -1,4 +1,8 @@
-﻿namespace Persons;
+﻿using TASK_4.Persons;
+using TASK_4.Reporters;
+using TASK_4.Utils;
+
+namespace TASK_4;
 
 public class Program
 {
@@ -9,11 +13,14 @@ public class Program
         var (candidates, employees) = FillInCreatedLists(emptyListOfCandidates, emptyListOfCEmployees);
 
         ShowCandidatesAndEmployeesInfo(candidates, employees);
-        
+        PressKeyToContinue();
+
         ReportOnCandidates(candidates);
+        PressKeyToContinue();
+
         ReportOnEmployees(employees);
     }
-    
+
     public static Tuple<List<Candidate>, List<Employee>> CreateEmptyListsOfCandidatesAndEmployees()
     {
         var listOfObjectsCandidates = new List<Candidate>();
@@ -34,7 +41,7 @@ public class Program
         List<Candidate> listOfObjectsCandidates,
         List<Employee> listOfObjectsEmployees)
     {
-        var repository = new PersonsRepository();
+        var repository = new BogusRepository();
         var candidates = repository.GetCandidates(listOfObjectsCandidates);
         var employees = repository.GetEmployees(listOfObjectsEmployees);
         Console.WriteLine();
@@ -45,14 +52,16 @@ public class Program
     public static void ShowCandidatesAndEmployeesInfo(IEnumerable<Candidate> candidates,
         IEnumerable<Employee> employees)
     {
-        candidates.ElementAt(0).ShowUserInfo();
-        employees.ElementAt(0).ShowUserInfo();
+        const int userNumberInTheList = 0;
+        candidates.ElementAt(userNumberInTheList).ShowUserInfo();
+        employees.ElementAt(userNumberInTheList).ShowUserInfo();
         Console.WriteLine();
     }
 
     public static void ReportOnEmployees(IEnumerable<Employee> employees)
     {
         var reportEmployees = new EmployeeReportGenerator(employees);
+        Console.WriteLine("List of employees:");
         reportEmployees.Report();
         Console.WriteLine();
     }
@@ -60,7 +69,14 @@ public class Program
     public static void ReportOnCandidates(IEnumerable<Candidate> candidates)
     {
         var reportCandidates = new CandidateReportGenerator(candidates);
+        Console.WriteLine("List of candidates:");
         reportCandidates.Report();
         Console.WriteLine();
+    }
+
+    public static void PressKeyToContinue()
+    {
+        Console.WriteLine("Press any key to continue.\n");
+        Console.ReadKey();
     }
 }
